Ezra_Medical_Chatbot/requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client
langchain
langchain_community
langchain_pinecone
flask
python-dotenv
-e .